refactor: HomeRepository 제거, StarRecordRepositoryPort로 아키텍처 경계 정리
